docker/services/nuclios-server/requirements.txt
aioredis==2.0.1
alembic==1.13.1
azure-cosmosdb-table==1.0.6
azure-storage-blob==1.4.0
azure-storage-file-share==12.15.0
beautifulsoup4==4.11.1
bidict==0.22.1
biopython==1.81
boto3==1.34.88
cryptography==42.0.8
dask==2022.1.1
dnspython==2.6.1
eventlet==0.36.1
fastapi==0.108.0
fastapi-socketio==0.0.10
flasgger==0.9.5
gevent==24.2.1
gmaps==0.9.0
google-api-core==2.17.0
google-auth==2.27.0
html-sanitizer==2.4.4
httpcore==1.0.2
httpx==0.25.2
hypercorn==0.17.3
ipython==8.21.0
ipywidgets==7.6.3
itsdangerous==2.0.1
json5==0.9.14
jupyterlab==4.1.1
jupyterlab-server==2.25.2
kaleido==0.2.1
kfp==2.8.0
langchain==0.2.14
lxml_html_clean==0.1.0
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.8.2
millify==0.1.1
mistune==0.8.4
msal==1.26.0
msrest==0.7.1
msrestazure==0.6.4.post1
natsort==8.4.0
networkx==3.1
notebook-shim==0.2.3
numpy==1.26.4
openai==1.30.1
opencensus-ext-azure==1.1.13
opencensus-ext-fastapi==0.1.0
opencensus-ext-flask==0.8.2
opencv-python-headless==4.9.0.80
openpyxl==3.0.7
pandas==1.5.3
parquet==1.3.1
passlib==1.7.4
plotly==5.21.0
prometheus-client==0.19.0
psutil==5.8.0
psycopg2-binary==2.9.5
pyarrow==17.0.0
pycountry==22.3.5
pydantic==2.4.2
pydantic-settings==2.0.3
PyJWT==2.7.0
pymongo==4.8.0
pyodbc==4.0.39
PyPDF2==3.0.1
pysaml2==7.5.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
scikit-learn==1.5.1
scipy == 1.10.1
sendgrid==6.9.7
sentry-sdk[fastapi]==2.13.0
simple-websocket
snowflake-connector-python==3.11.0
SQLAlchemy==1.4.46
starlette==0.32.0
statsmodels == 0.13.5
threadpoolctl==3.2.0
uvicorn==0.23.2
Werkzeug==3.0.4
wordcloud==1.9.2
xlrd==2.0.1
